requirements.txt
PyGithub>=1.58.0
GitPython>=3.1.30